fix: using next script for gtag and removing custom document
