+++ pages/_app.tsx
@@ -2,6 +2,7 @@ import ProgressBar from '@badrap/bar-of-progress'
 
 import Router, { useRouter } from 'next/router'
 import type { AppProps } from 'next/app'
+import Script from 'next/script'
 
 import Layout from 'components/layouts/Layout'
 
@@ -43,6 +44,26 @@ function MyApp({ Component, pageProps }: AppProps) {
 
 	return (
 		<>
+			{/* Global Site Tag (gtag.js) - Google Analytics */}
+			<Script
+				strategy='afterInteractive'
+				src={`https://www.googletagmanager.com/gtag/js?id=${gtag.GA_TRACKING_ID}`}
+			/>
+			<Script
+				id='gtag-init'
+				strategy='afterInteractive'
+				dangerouslySetInnerHTML={{
+					__html: `
+            window.dataLayer = window.dataLayer || [];
+            function gtag(){dataLayer.push(arguments);}
+            gtag('js', new Date());
+            gtag('config', '${gtag.GA_TRACKING_ID}', {
+              page_path: window.location.pathname,
+            });
+          `,
+				}}
+			/>
+
 			{isLoginPage ? (
 				<Component {...pageProps} />
 			) : (
