--- pages/_document.tsx
@@ -1,37 +1,0 @@
-/**
- * Custom document, see: https://nextjs.org/docs/advanced-features/custom-document
- */
-
-import Document, { Html, Head, Main, NextScript } from 'next/document'
-
-export default class MyDocument extends Document {
-	render() {
-		return (
-			<Html>
-				<Head>
-					{/* Global Site Tag (gtag.js) - Google Analytics */}
-					<script
-						async
-						src={`https://www.googletagmanager.com/gtag/js?id=${process.env.FIREBASE_MEASUREMENT_ID}`}
-					/>
-					<script
-						dangerouslySetInnerHTML={{
-							__html: `
-            window.dataLayer = window.dataLayer || [];
-            function gtag(){dataLayer.push(arguments);}
-            gtag('js', new Date());
-            gtag('config', '${process.env.FIREBASE_MEASUREMENT_ID}', {
-              page_path: window.location.pathname,
-            });
-          `,
-						}}
-					/>
-				</Head>
-				<body>
-					<Main />
-					<NextScript />
-				</body>
-			</Html>
-		)
-	}
-}
